Chore: Working project

+++ server/bin/commands/dev/dev.ts
@@ -1,8 +1,8 @@
 import { Command } from "commander";
 import { terminal } from "../../../src/main";
 import getConfig from "../../utils/getConfig";
-import { spawn, ChildProcess } from 'child_process';
-import path from 'path';
+import { spawn, ChildProcess } from "child_process";
+import path from "path";
 import chokidar, { FSWatcher } from "chokidar";
 import fs from "fs-extra";
 
@@ -18,7 +18,9 @@ interface CommandOptions {
  */
 function exitApp() {
     terminal.error("The dev server failed to start");
-    terminal.error("Check debug log at (__DOES_NOT_EXIST__) for detailed information");
+    terminal.error(
+        "Check debug log at (__DOES_NOT_EXIST__) for detailed information"
+    );
     process.exit();
 }
 
@@ -27,27 +29,52 @@ function exitApp() {
  * @param bin Commander program
  */
 export default function dev(bin: Command) {
-    bin
-        .command("dev [path]")
-        .option("--config <path>", "The path to your application's configuration file")
+    bin.command("dev [path]")
+        .option(
+            "--config <path>",
+            "The path to your application's configuration file"
+        )
         .description("Start a development server")
         .action((pathName = "./", options: CommandOptions) => {
             getConfig(options.config).then((config) => {
-                terminal.info(`The app will now boot in development for ${config.app.type} platforms`);
-                terminal.info("The renderer process is starting in development");
+                terminal.info(
+                    `The app will now boot in development for ${config.app.type} platforms`
+                );
+                terminal.info(
+                    "The renderer process is starting in development"
+                );
 
                 if (config.app.type == "desktop") {
-                    startDesktopDevServer(path.join(process.cwd(), pathName)).then((port) => {
-                        terminal.success("The development server has loaded on port " + port);
-                        terminal.info("The desktop process is starting in development");
+                    startDesktopDevServer(
+                        path.join(process.cwd(), pathName)
+                    ).then((port) => {
+                        terminal.success(
+                            "The development server has loaded on port " + port
+                        );
+                        terminal.info(
+                            "The desktop process is starting in development"
+                        );
 
-                        startDesktopDevApp(path.join(process.cwd(), pathName), port, [ ...config.app.desktop?.externalDirs!, config.app.desktop?.electronRoot! ]).then(() => {
-                            terminal.success("The development app has been loaded");
-                            terminal.success("App is fully ready for being developed");
+                        startDesktopDevApp(
+                            path.join(process.cwd(), pathName),
+                            port,
+                            [
+                                ...config.app.desktop?.externalDirs!,
+                                config.app.desktop?.electronRoot!,
+                            ]
+                        ).then(() => {
+                            terminal.success(
+                                "The development app has been loaded"
+                            );
+                            terminal.success(
+                                "App is fully ready for being developed"
+                            );
                         });
                     });
                 } else {
-                    terminal.error("EUNIMP: Cannot deploy mobile applications on this version of LuxJS. Either update your LuxJS service or wait for an update");
+                    terminal.error(
+                        "EUNIMP: Cannot deploy mobile applications on this version of LuxJS. Either update your LuxJS service or wait for an update"
+                    );
                 }
             });
         });
@@ -60,9 +87,13 @@ export default function dev(bin: Command) {
  */
 function startDesktopDevServer(rootPath: string): Promise<number> {
     return new Promise((resolve, reject) => {
-        const devServer = spawn("node", [ "./subProc/desktop/vite.js", rootPath ], {
-            cwd: __dirname
-        });
+        const devServer = spawn(
+            "node",
+            ["./subProc/desktop/vite.js", rootPath],
+            {
+                cwd: __dirname,
+            }
+        );
 
         let devServerListening = false;
 
@@ -83,26 +114,34 @@ function startDesktopDevServer(rootPath: string): Promise<number> {
 /**
  * Start the ElectronJS app in dev
  * @param rootPath The app root
- * @param port Port of the dev server 
+ * @param port Port of the dev server
  * @returns A promise for when the app has been started
  */
-function startDesktopDevApp(rootPath: string, serverPort: number, electronDirs: string[]): Promise<void> {
+function startDesktopDevApp(
+    rootPath: string,
+    serverPort: number,
+    electronDirs: string[]
+): Promise<void> {
     return new Promise((resolve, reject) => {
         let devApp: ChildProcess;
         let ready = false;
 
         const startDevApp = () => {
-            devApp = spawn("node", [ "./subProc/desktop/electron.js", rootPath, serverPort + "" ], {
-                cwd: __dirname
-            });
-    
+            devApp = spawn(
+                "node",
+                ["./subProc/desktop/electron.js", rootPath, serverPort + ""],
+                {
+                    cwd: __dirname,
+                }
+            );
+
             devApp.stdout!.on("data", (d) => {
                 const data = d.toString().split("\n") as string[];
-    
-                data.forEach(line => {
+
+                data.forEach((line) => {
                     try {
                         const message = JSON.parse(line) as any;
-    
+
                         if (!ready && message.boot && message.boot.done) {
                             ready = true;
                             resolve();
@@ -113,12 +152,17 @@ function startDesktopDevApp(rootPath: string, serverPort: number, electronDirs: 
 
             devApp.stderr?.on("data", (d) => {
                 const data = d.toString().split("\n") as string[];
+                terminal.error("An error occurred in the electron app");
 
-                data.forEach(line => {
+                data.forEach((line) => {
                     terminal.error("[ Electron ] " + line);
                 });
-            })
-        }
+
+                terminal.error(
+                    "An error occurred in the electron app, please read above for output logging"
+                );
+            });
+        };
 
         startDevApp();
         const watchers = [] as FSWatcher[];
@@ -129,9 +173,13 @@ function startDesktopDevApp(rootPath: string, serverPort: number, electronDirs: 
                     watchers.push(chokidar.watch(path.join(rootPath, dir)));
                 } else {
                     if (electronDirs.length == index + 1) {
-                        terminal.error("The electron's root path is a file but should be a directory");
+                        terminal.error(
+                            "The electron's root path is a file but should be a directory"
+                        );
                     } else {
-                        terminal.error("One or more of the electron external paths route to a file but should go to a directory instead");
+                        terminal.error(
+                            "One or more of the electron external paths route to a file but should go to a directory instead"
+                        );
                     }
 
                     exitApp();
@@ -140,17 +188,26 @@ function startDesktopDevApp(rootPath: string, serverPort: number, electronDirs: 
                 if (electronDirs.length == index + 1) {
                     terminal.error("The electron's root path is invalid");
                 } else {
-                    terminal.error("One or more of the electron external paths are invalid");
+                    terminal.error(
+                        "One or more of the electron external paths are invalid"
+                    );
                 }
 
                 exitApp();
             }
         });
 
-        watchers.forEach(watcher => {
+        watchers.forEach((watcher) => {
             watcher.on("all", () => {
-                devApp?.kill("SIGTERM");
-                startDevApp();
+                if (ready) {
+                    terminal.info("The desktop app is reloading");
+
+                    devApp.on("exit", () => {
+                        startDevApp();
+                    });
+
+                    devApp?.kill("SIGTERM");
+                }
             });
         });
     });
